회원가입

+++ src/main/java/org/choongang/member/mapper/MemberMapper.java
@@ -3,7 +3,7 @@ package org.choongang.member.mapper;
 import org.choongang.member.entities.Member;
 
 public interface MemberMapper {
-    long exist(String email);
+    int exist(String email);
     Member get(String email);
     int register(Member member);
